feature: js interface for request permission
